+++ experiments/les_strats/tc_calibrate.jl
@@ -5,6 +5,9 @@
 @everywhere using Distributions
 @everywhere using StatsBase
 @everywhere using LinearAlgebra
+@everywhere using CalibrateEDMF.ReferenceModels
+@everywhere using CalibrateEDMF.ReferenceStats
+@everywhere using CalibrateEDMF.TurbulenceConvectionUtils
 # Import EKP modules
 @everywhere using EnsembleKalmanProcesses.EnsembleKalmanProcessModule
 @everywhere using EnsembleKalmanProcesses.Observations
