+++ experiments/scm_pycles_pipeline/hpc_parallel/init.jl
@@ -3,6 +3,9 @@
 using Distributions
 using StatsBase
 using LinearAlgebra
+using CalibrateEDMF.ReferenceModels
+using CalibrateEDMF.ReferenceStats
+using CalibrateEDMF.TurbulenceConvectionUtils
 # Import EKP modules
 using EnsembleKalmanProcesses.EnsembleKalmanProcessModule
 using EnsembleKalmanProcesses.Observations
