+++ experiments/scm_pycles_pipeline/hpc_parallel/single_scm_eval.jl
@@ -2,6 +2,9 @@
 #
 using ArgParse
 using Distributions
+using CalibrateEDMF.ReferenceModels
+using CalibrateEDMF.ReferenceStats
+using CalibrateEDMF.TurbulenceConvectionUtils
 # Import EKP modules
 using EnsembleKalmanProcesses.EnsembleKalmanProcessModule
 using EnsembleKalmanProcesses.Observations
