+++ src/helper_funcs.jl
@@ -5,282 +5,13 @@ using LinearAlgebra
 using Glob
 using JSON
 using Random
+using ..ReferenceModels
 # EKP modules
 using EnsembleKalmanProcesses.ParameterDistributionStorage
 # TurbulenceConvection.jl
 using TurbulenceConvection
 tc_dir = dirname(dirname(pathof(TurbulenceConvection)));
 include(joinpath(tc_dir, "integration_tests", "utils", "main.jl"))
-
-"""
-    struct ReferenceModel
-    
-A structure containing information about the 'true' reference model
-and the observation map used to compare the parameterized and
-reference models.
-"""
-Base.@kwdef struct ReferenceModel
-    "Vector of reference variable names"
-    y_names::Vector{String}
-
-    "Root directory for reference LES data"
-    les_root::String
-    "Name of LES reference simulation file"
-    les_name::String
-    "Suffix of LES reference simulation file"
-    les_suffix::String
-
-    "Root directory for SCM data (used for interpolation)"
-    scm_root::String
-    "Name of SCM reference simulation file"
-    scm_name::String
-    "Suffix of SCM reference simulation file"
-    scm_suffix::String = "00000"
-
-    # TODO: Make t_start and t_end vectors for multiple time intervals per reference model.
-    "Start time for computing statistics over"
-    t_start::Real
-    "End time for computing statistics over"
-    t_end::Real
-end
-
-les_dir(m::ReferenceModel) = data_directory(m.les_root, m.les_name, m.les_suffix)
-scm_dir(m::ReferenceModel) = data_directory(m.scm_root, m.scm_name, m.scm_suffix)
-data_directory(root::S, name::S, suffix::S) where {S <: AbstractString} = joinpath(root, "Output.$name.$suffix")
-
-namelist_directory(root::String, m::ReferenceModel) = namelist_directory(root, m.scm_name)
-namelist_directory(root::S, casename::S) where {S <: AbstractString} = joinpath(root, "namelist_$casename.in")
-
-num_vars(m::ReferenceModel) = length(m.y_names)
-
-
-"""
-    struct ReferenceStatistics
-    
-A structure containing statistics from the reference model used to
-define a well-posed inverse problem.
-"""
-Base.@kwdef struct ReferenceStatistics{FT <: Real}
-    "Reference data, length: nSim * n_vars * n_zLevels(possibly reduced by PCA)"
-    y::Vector{FT} # yt
-    "Data covariance matrix, dims: (y,y) (possibly reduced by PCA)"
-    Γ::Array{FT, 2}  # Γy
-    "Vector (length: nSim) of normalizing factors (length: n_vars)"
-    norm_vec::Vector{Array{FT, 1}}  # pool_var_list
-
-    "Vector (length: nSim) of PCA projection matrices with leading eigenvectors as columns"
-    pca_vec::Vector{Union{Array{FT, 2}, UniformScaling}}  # P_pca_list
-
-    "Full reference data vector, length: nSim * n_vars * n_zLevels"
-    y_full::Vector{FT}  # yt_big
-    "Full covariance matrix, dims: (y,y)"
-    Γ_full::Array{FT, 2}  # yt_var_big
-
-    """
-        ReferenceStatistics(
-            RM::Vector{ReferenceModel},
-            model_type::Symbol,
-            perform_PCA::Bool,
-            normalize::Bool,
-            FT::DataType = Float64;
-            variance_loss::Float64 = 0.1,
-            tikhonov_noise::Float64 = 0.0,
-            tikhonov_mode::String = "absolute",
-            Γ_scaling::Float64 = 1.0,
-        )
-
-    Constructs the ReferenceStatistics defining the inverse problem.
-
-    Inputs:
-     - RM               :: Vector of `ReferenceModel`s
-     - model_type       :: Type of the reference model, either :les or :scm.
-     - perform_PCA      :: Boolean specifying whether to perform PCA.
-     - normalize        :: Boolean specifying whether to normalize the data.
-     - variance_loss    :: Fraction of variance loss when performing PCA.
-     - tikhonov_noise   :: Tikhonov regularization factor for covariance matrices.
-     - tikhonov_mode    :: If "relative", tikhonov_noise is scaled by the minimum
-        eigenvalue in the covariance matrix considered.
-     - dim_scaling      :: Whether to scale covariance blocks by their size.
-    Outputs:
-     - A ReferenceStatistics struct.
-    """
-    function ReferenceStatistics(
-        RM::Vector{ReferenceModel},
-        model_type::Symbol,
-        perform_PCA::Bool,
-        normalize::Bool,
-        FT::DataType = Float64;
-        variance_loss::Float64 = 0.1,
-        tikhonov_noise::Float64 = 0.0,
-        tikhonov_mode::String = "absolute",
-        dim_scaling::Bool = false,
-    )
-        # Init arrays
-        y = FT[]  # yt
-        Γ_vec = Array{FT, 2}[]  # yt_var_list
-        y_full = FT[]  # yt_big
-        Γ_full_vec = Array{FT, 2}[]  # yt_var_list_big
-        pca_vec = []  # P_pca_list
-        norm_vec = Vector[]  # pool_var_list
-
-        for m in RM
-            # Get (interpolated and pool-normalized) observations, get pool variance vector
-            y_, y_var_, pool_var = get_obs(model_type, m, z_scm = get_height(scm_dir(m)), normalize)
-            push!(norm_vec, pool_var)
-            if perform_PCA
-                y_pca, y_var_pca, P_pca = obs_PCA(y_, y_var_, variance_loss)
-                append!(y, y_pca)
-                push!(Γ_vec, y_var_pca)
-                push!(pca_vec, P_pca)
-            else
-                append!(y, y_)
-                push!(Γ_vec, y_var_)
-                push!(pca_vec, 1.0I)
-            end
-            # Save full dimensionality (normalized) output for error computation
-            append!(y_full, y_)
-            push!(Γ_full_vec, y_var_)
-        end
-
-        # Construct global observational covariance matrix, original space
-        Γ_full = cat(Γ_full_vec..., dims = (1, 2))
-
-        # Construct global observational covariance matrix, TSVD
-        if tikhonov_mode == "relative"
-            Γ_vec = map(x -> x + tikhonov_noise * minimum(diag(x)) * I, Γ_vec)
-        else
-            Γ_vec = map(x -> x + tikhonov_noise * I, Γ_vec)
-        end
-        # Scale by number of dimensions
-        Γ_vec = dim_scaling ? map(x -> size(x, 1) * x, Γ_vec) : Γ_vec
-        Γ = cat(Γ_vec..., dims = (1, 2))
-        @assert isposdef(Γ)
-
-        return new{FT}(y, Γ, norm_vec, pca_vec, y_full, Γ_full)
-    end
-
-    ReferenceStatistics(y::Vector{FT}, args...) where {FT <: Real} = new{FT}(y, args...)
-end
-
-pca_length(RS::ReferenceStatistics) = length(RS.y)
-full_length(RS::ReferenceStatistics) = length(RS.y_full)
-
-
-"""
-    run_SCM(
-        u::Vector{FT},
-        u_names::Vector{String},
-        RM::Vector{ReferenceModel},
-        RS::ReferenceStatistics,
-    ) where FT<:Real
-
-Run the single-column model (SCM) using a set of parameters u 
-and return the value of outputs defined in y_names, possibly 
-after normalization and projection onto lower dimensional 
-space using PCA.
-
-Inputs:
- - u                :: Values of parameters to be used in simulations.
- - u_names          :: SCM names for parameters `u`.
- - RM               :: Vector of `ReferenceModel`s
- - RS               :: reference statistics for simulation
-Outputs:
- - sim_dirs         :: Vector of simulation output directories
- - g_scm            :: Vector of model evaluations concatenated for all flow configurations.
- - g_scm_pca        :: Projection of `g_scm` onto principal subspace spanned by eigenvectors.
-"""
-function run_SCM(
-    u::Vector{FT},
-    u_names::Vector{String},
-    RM::Vector{ReferenceModel},
-    RS::ReferenceStatistics,
-) where {FT <: Real}
-
-    g_scm = zeros(0)
-    g_scm_pca = zeros(0)
-    sim_dirs = String[]
-
-    mkpath(joinpath(pwd(), "tmp"))
-    for (i, m) in enumerate(RM)
-        # create temporary directory to store SCM data in
-        tmpdir = mktempdir(joinpath(pwd(), "tmp"))
-
-        # run TurbulenceConvection.jl. Get output directory for simulation data
-        sim_dir = run_SCM_handler(m, tmpdir, u, u_names)
-        push!(sim_dirs, sim_dir)
-
-        g_scm_flow = get_profile(m, sim_dir)
-        # normalize
-        g_scm_flow = normalize_profile(g_scm_flow, length(m.y_names), RS.norm_vec[i])
-        append!(g_scm, g_scm_flow)
-
-        # perform PCA reduction
-        append!(g_scm_pca, RS.pca_vec[i]' * g_scm_flow)
-    end
-
-    # penalize nan-values in output
-    any(isnan.(g_scm)) && warn("NaN-values in output data")
-    g_scm[isnan.(g_scm)] .= 1e5
-
-    g_scm_pca[isnan.(g_scm_pca)] .= 1e5
-    println("LENGTH OF G_SCM_ARR : ", length(g_scm))
-    println("LENGTH OF G_SCM_ARR_PCA : ", length(g_scm_pca))
-    return sim_dirs, g_scm, g_scm_pca
-end
-
-
-"""
-    run_SCM_handler(
-        m::ReferenceModel,
-        tmpdir::String,
-        u::Array{FT, 1},
-        u_names::Array{String, 1},
-    ) where {FT<:AbstractFloat}
-
-Run a list of cases using a set of parameters `u_names` with values `u`,
-and return a list of directories pointing to where data is stored for 
-each simulation run.
-
-Inputs:
- - m            :: Reference model
- - tmpdir       :: Directory to store simulation results in
- - u            :: Values of parameters to be used in simulations.
- - u_names      :: SCM names for parameters `u`.
-Outputs:
- - output_dirs  :: list of directories containing output data from the SCM runs.
-"""
-function run_SCM_handler(
-    m::ReferenceModel,
-    tmpdir::String,
-    u::Array{FT, 1},
-    u_names::Array{String, 1},
-) where {FT <: AbstractFloat}
-
-    # fetch default namelist
-    inputdir = scm_dir(m)
-    namelist = JSON.parsefile(namelist_directory(inputdir, m))
-
-    # update parameter values
-    for (pName, pVal) in zip(u_names, u)
-        namelist["turbulence"]["EDMF_PrognosticTKE"][pName] = pVal
-    end
-
-    # set random uuid
-    uuid = basename(tmpdir)
-    namelist["meta"]["uuid"] = uuid
-    # set output dir to `tmpdir`
-    namelist["output"]["output_root"] = tmpdir
-    # write updated namelist to `tmpdir`
-    namelist_path = namelist_directory(tmpdir, m)
-    open(namelist_path, "w") do io
-        JSON.print(io, namelist, 4)
-    end
-
-    # run TurbulenceConvection.jl with modified parameters
-    main(namelist)
-
-    return data_directory(tmpdir, m.scm_name, uuid)
-end
 
 
 """
@@ -682,30 +413,6 @@ Deserializes the given dictionary and constructs a struct
 of the given type with the dictionary values.
 """
 deserialize_struct(dict::Dict{String}, ::Type{T}) where {T} = T(map(fn -> dict["$fn"], fieldnames(T))...)
-
-"""
-    generate_scm_input(u::Vector{Float64},
-        u_names::Vector{String},
-        RM::Vector{ReferenceModel},
-        RS::ReferenceStatistics,)
-
-Generates all data necessary to initialize a SCM evaluation
-at the given parameter vector `u`.
-"""
-function generate_scm_input(
-    u::Vector{Float64},
-    u_names::Vector{String},
-    RM::Vector{ReferenceModel},
-    RS::ReferenceStatistics,
-    outdir_path::String = pwd(),
-)
-    # Generate version
-    version = rand(11111:99999)
-    ref_models = map(x -> serialize_struct(x), RM)
-    ref_stats = serialize_struct(RS)
-    jldsave(scm_init_path(outdir_path, version); u, u_names, ref_models, ref_stats, version)
-    return version
-end
 
 """
     jld2_path(root::String, identifier::Union{String, Int}, prefix::String)
